@@ -4,28 +4,28 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
-  flutter: "3.24.1"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: "3.29.2"
 
 dependencies:
-  bloc_concurrency: ^0.2.5
+  bloc_concurrency: 0.3.0
   core_design:
     path: ../../../packages/core/design
-  dart_mappable: ^4.2.2
+  dart_mappable: 4.5.0
   flutter:
     sdk: flutter
-  google_mobile_ads: 5.2.0
-  injectable: ^2.3.2
-  json_annotation: ^4.8.1
+  google_mobile_ads: ^5.3.1
+  injectable: 2.5.0
+  json_annotation: 4.9.0
   utility_di:
     path: ../../../packages/utility/di
   utility_navigation:
     path: ../../../packages/utility/navigation
 
 dev_dependencies:
-  build_runner: ^2.4.9
-  dart_mappable_builder: ^4.2.3
-  injectable_generator: ^2.4.2
+  build_runner: 2.4.15
+  dart_mappable_builder: 4.5.0
+  injectable_generator: 2.7.0
   utility_analysis:
     path: ../../../packages/utility/analysis
 
