@@ -4,26 +4,26 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
-  flutter: "3.24.1"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: "3.29.2"
 
 dependencies:
-  dio: ^5.4.3+1
-  dio_cache_interceptor: ^3.5.0
-  dio_cache_interceptor_hive_store: ^3.2.2
+  dio: 5.8.0+1
+  dio_cache_interceptor: 4.0.2
+  http_cache_hive_store: 5.0.1
+  hive: ^2.2.3
   flutter:
     sdk: flutter
-  hive: ^2.2.3
-  injectable: ^2.3.2
-  supabase_flutter: ^2.5.6
+  injectable: 2.5.0
+  supabase_flutter: 2.8.4
   utility_di:
     path: ../../../packages/utility/di
 
 dev_dependencies:
-  build_runner: ^2.4.9
+  build_runner: 2.4.15
   flutter_test:
     sdk: flutter
-  injectable_generator: ^2.4.2
+  injectable_generator: 2.7.0
   utility_analysis:
     path: ../../../packages/utility/analysis
 
