@@ -4,31 +4,31 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
-  flutter: "3.24.1"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: "3.29.2"
 
 dependencies:
   core_api_client:
     path: ../../../packages/core/api_client
   core_domain:
     path: ../../../packages/core/domain
-  cached_network_image: ^3.3.1
-  fl_chart: ^0.68.0
+  cached_network_image: 3.4.1
+  fl_chart: 0.70.2
   flutter:
     sdk: flutter
-  flutter_svg: ^2.0.10+1
-  google_fonts: ^6.2.0
+  flutter_svg: 2.0.17
+  google_fonts: 6.2.1
   intersperse: ^2.0.0
-  intl: ^0.19.0
+  intl: 0.19.0
   modal_bottom_sheet: ^3.0.0
   shimmer: ^3.0.0
-  stream_transform: ^2.1.0
+  stream_transform: 2.1.1
   utility_extensions:
     path: ../../../packages/utility/extensions
 
 dev_dependencies:
-  build_runner: ^2.4.9
-  flutter_gen_runner: ^5.4.0
+  build_runner: 2.4.15
+  flutter_gen_runner: 5.10.0
   utility_analysis:
     path: ../../../packages/utility/analysis
 
