@@ -4,38 +4,38 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
-  flutter: "3.24.1"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: "3.29.2"
 
 dependencies:
-  cached_annotation: ^1.7.0
-  collection: ^1.18.0
+  cached_annotation: 1.7.1
+  collection: 1.19.1
   core_analytics:
     path: ../../../packages/core/analytics
   core_api_client:
     path: ../../../packages/core/api_client
-  dart_mappable: ^4.2.2
-  dio: ^5.4.3+1
+  dart_mappable: 4.5.0
+  dio: 5.8.0+1
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.5
+  flutter_bloc: 9.1.0
   hive: ^2.2.3
-  injectable: ^2.3.2
-  json_annotation: ^4.8.1
-  retrofit: ^4.4.1
-  supabase_flutter: ^2.5.6
+  injectable: 2.5.0
+  json_annotation: 4.9.0
+  retrofit: 4.4.2
+  supabase_flutter: 2.8.4
   utility_di:
     path: ../../../packages/utility/di
   utility_extensions:
     path: ../../../packages/utility/extensions
 
 dev_dependencies:
-  build_runner: ^2.4.9
-  cached: 1.7.0
-  dart_mappable_builder: ^4.2.3
-  injectable_generator: ^2.4.2
-  json_serializable: ^6.7.1
-  retrofit_generator: ^9.1.5
+  build_runner: 2.4.15
+  cached: 1.7.1
+  dart_mappable_builder: 4.5.0
+  injectable_generator: 2.7.0
+  json_serializable: 6.9.4
+  retrofit_generator: 9.1.9
   utility_analysis:
     path: ../../../packages/utility/analysis
 
