@@ -4,11 +4,11 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
-  flutter: "3.24.1"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: "3.29.2"
 
 dependencies:
-  bloc_concurrency: ^0.2.5
+  bloc_concurrency: 0.3.0
   core_ads:
     path: ../../../packages/core/ads
   core_analytics:
@@ -19,22 +19,22 @@ dependencies:
     path: ../../../packages/core/design
   core_domain:
     path: ../../../packages/core/domain
-  dart_mappable: ^4.2.2
+  dart_mappable: 4.5.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.5
-  injectable: ^2.3.2
-  json_annotation: ^4.8.1
+  flutter_bloc: 9.1.0
+  injectable: 2.5.0
+  json_annotation: 4.9.0
   utility_di:
     path: ../../../packages/utility/di
   utility_navigation:
     path: ../../../packages/utility/navigation
 
 dev_dependencies:
-  build_runner: ^2.4.9
-  dart_mappable_builder: ^4.2.3
-  injectable_generator: ^2.4.2
-  json_serializable: ^6.7.1
+  build_runner: 2.4.15
+  dart_mappable_builder: 4.5.0
+  injectable_generator: 2.7.0
+  json_serializable: 6.9.4
   utility_analysis:
     path: ../../../packages/utility/analysis
 
