@@ -42,7 +42,7 @@ class AttributesCard extends StatelessWidget {
       child: DecoratedBox(
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(AppCornerRadius.radius2),
-          color: context.colors.contentSecondary.withOpacity(0.2),
+          color: context.colors.contentSecondary.withAlpha(50),
         ),
         child: Padding(
           padding: const EdgeInsets.symmetric(
