@@ -4,5 +4,5 @@ export 'package:core_design/src/organisms/organisms.dart';
 export 'package:core_design/src/templates/templates.dart';
 export 'package:core_design/src/utility/app_formatters.dart';
 export 'package:core_design/src/utility/get_player_colors.dart';
+export 'package:core_design/src/utility/x_int.dart';
 export 'package:intersperse/intersperse.dart';
-export 'package:fl_chart/fl_chart.dart';
